eliminazione commenti

+++ src/main/java/it/pn/frontend/e2e/common/HelpdeskPage.java
@@ -112,20 +112,11 @@ public class HelpdeskPage extends BasePage {
             getWebDriverWait(40).withMessage("password non presente").until(ExpectedConditions.visibilityOf(passwordInput));
             getWebDriverWait(40).withMessage("submit non presente").until(ExpectedConditions.visibilityOf(loginButton));
         } catch (TimeoutException e) {
-            logger.error("Form non presente errore: " + e.getMessage());
             Assertions.fail("Form non presente errore: " + e.getMessage());
         }
     }
 
     public void checkHome() {
-//        try {
-//            monitoraggioPN = driver.findElement(By.id("cardTitle-Monitoraggio Piattaforma Notifiche"));
-//            getWebDriverWait(10).withMessage("home non presente").until(ExpectedConditions.visibilityOf(monitoraggioPN));
-//            logger.info("pagina home carica");
-//        } catch (TimeoutException e) {
-//            logger.error("errore caricamento home helpdesk: " + e.getMessage());
-//            Assertions.fail("errore caricamento home helpdesk: " + e.getMessage());
-//        }
         try {
             logger.info("In attesa che la pagina home venga caricata...");
 
@@ -139,12 +130,6 @@ public class HelpdeskPage extends BasePage {
     }
 
     public void waitLoadServiceTable() {
-//        services = driver.findElements(By.xpath(".//div[@data-field='functionality' and @role='cell']"));
-//        if (services.size() < 3) {
-//            Assertions.fail("I servizi visualizzati sono meno di 3");
-//        }
-//        getWebDriverWait(10).withMessage("Non è visibile la tabella dei disservizi").until(ExpectedConditions.visibilityOfAllElements(services));
-//        By serviceXpath = By.xpath(".//div[@data-field='functionality' and @role='cell']");
 
         List<WebElement> services = getWebDriverWait(10)
                 .withMessage("Non è visibile la tabella dei disservizi o i servizi sono meno di 3")
@@ -181,15 +166,6 @@ public class HelpdeskPage extends BasePage {
     }
 
     public void clickMonitoraggio() {
-//        try {
-//            logger.info("clicco sulla card monitoraggio piattaforma notifiche");
-//            monitoraggioPN = driver.findElement(By.id("cardTitle-Monitoraggio Piattaforma Notifiche"));
-//            getWebDriverWait(10).withMessage("Il bottone monitoraggio non è cliccabile").until(ExpectedConditions.elementToBeClickable(monitoraggioPN));
-//            monitoraggioPN.click();
-//        } catch (TimeoutException e) {
-//            logger.error("Card monitoraggio non cliccabile: " + e.getMessage());
-//            Assertions.fail("Card monitoraggio non cliccabile: " + e.getMessage());
-//        }
 
         try {
             logger.info("In attesa che la card 'Monitoraggio Piattaforma Notifiche' sia cliccabile...");
@@ -317,18 +293,6 @@ public class HelpdeskPage extends BasePage {
     }
 
     public boolean checkIsCreatedDisservizio() {
-//        try {
-//            WebElement dateDisservizio = elements(By.xpath("//div[@data-field='data']")).get(1);
-//            getWebDriverWait(10).until(ExpectedConditions.visibilityOf(dateDisservizio));
-//            if (dateDisservizio.getText() != null && !dateDisservizio.getText().isEmpty()) {
-//                logger.info("disservizio già in corso");
-//                return true;
-//            }
-//            return false;
-//        } catch (TimeoutException e) {
-//            Assertions.fail("disservizio non creato: " + e.getMessage());
-//            return false;
-//        }
         try {
             List<WebElement> dateElements = getWebDriverWait(10)
                     .until(ExpectedConditions.visibilityOfAllElementsLocatedBy(By.xpath("//div[@data-field='data']")));
@@ -376,7 +340,6 @@ public class HelpdeskPage extends BasePage {
             getWebDriverWait(30).withMessage("button ricerca non trovato").until(ExpectedConditions.visibilityOf(buttonRicerca));
             getWebDriverWait(30).withMessage("button reset filtri non trovato").until(ExpectedConditions.visibilityOf(buttonResetFiltri));
         } catch (TimeoutException e) {
-            logger.error("home ricerca non caricata correttamente: " + e.getMessage());
             Assertions.fail("home ricerca non caricata correttamente: " + e.getMessage());
         }
     }
@@ -395,7 +358,6 @@ public class HelpdeskPage extends BasePage {
             logger.info("click conferma logout");
             elements(buttonConfermaLogout).get(0).click();
         } catch (TimeoutException e) {
-            logger.error("logout non riuscito correttamente: " + e.getMessage());
             Assertions.fail("logout non riuscito correttamente: " + e.getMessage());
         }
     }
@@ -505,7 +467,6 @@ public class HelpdeskPage extends BasePage {
             By checkUidIsDisplayed = By.id("Codice Univoco (uid)");
             this.getWebDriverWait(30).withMessage("input uid non trovato").until(ExpectedConditions.visibilityOfElementLocated(checkUidIsDisplayed));
         } catch (TimeoutException e) {
-            logger.error("opzione ottieni cf non trovata: " + e.getMessage());
             Assertions.fail("opzione ottieni cf non trovata: " + e.getMessage());
         }
     }
@@ -550,7 +511,6 @@ public class HelpdeskPage extends BasePage {
             By messaggio = By.xpath("//p[contains(text(),'intervallo temporale non può superare i 3 mesi')]");
             getWebDriverWait(15).withMessage("Messaggio di errore non trovato").until(ExpectedConditions.visibilityOfElementLocated(messaggio));
         } catch (TimeoutException e) {
-            logger.error("Messaggio di errore non trovato: " + e.getMessage());
             Assertions.fail("Messaggio di errore non trovato: " + e.getMessage());
         }
     }
@@ -592,49 +552,6 @@ public class HelpdeskPage extends BasePage {
             FileUtils.copyURLToFile(new URL(url), destFile, 5000, 5000);
 
 
-//        if (!headless) {
-//            logger.info("controllo esistenza link per scaricare zip");
-//            By zipLink = By.xpath("//a[contains(text(),'Download')]");
-//            WebElement linkZip = getWebDriverWait(10)
-//                    .withMessage("Link per scaricare zip non trovato")
-//                    .until(ExpectedConditions.visibilityOfElementLocated(zipLink));
-//            linkZip.click();
-//            Robot robot = new Robot();
-//            robot.setAutoDelay(100);
-//            robot.delay(2000);
-//            String workingDirectory = System.getProperty("user.dir");
-//            String path = workingDirectory + "/src/test/resources/dataPopulation/zip";
-//
-//            pressTabKey(robot, 6);
-//            robot.keyPress(KeyEvent.VK_ENTER);
-//            robot.keyRelease(KeyEvent.VK_ENTER);
-//
-//            typeFilePath(robot, path);
-//
-//            robot.keyPress(KeyEvent.VK_ENTER);
-//            robot.keyRelease(KeyEvent.VK_ENTER);
-//
-//            robot.delay(1000);
-//
-//            pressTabKey(robot, 8);
-//
-//            robot.keyPress(KeyEvent.VK_ENTER);
-//            robot.keyRelease(KeyEvent.VK_ENTER);
-//
-//            logger.info("Zip scaricato");
-//        } else {
-//            By zipLink = By.xpath("//a[contains(text(),'Download')]");
-//            String url = this.element(zipLink).getAttribute("href");
-//            String workingDirectory = System.getProperty("user.dir");
-//            File downloadDirectory = new File(workingDirectory + "/src/test/resources/dataPopulation/zip");
-//
-//            // Generate a unique filename for the downloaded ZIP file
-//            String fileName = "downloaded_" + System.currentTimeMillis() + ".zip";
-//
-//            File downloadFile = new File(downloadDirectory, fileName);
-//            FileUtils.copyURLToFile(new URL(url), downloadFile, 1000, 1000);
-//            logger.info("ZIP file downloaded successfully.");
-//        }
     }
 
     private void typeFilePath(Robot robot, String filePath) {
@@ -711,11 +628,6 @@ public class HelpdeskPage extends BasePage {
 
     public void checkPassword() {
         logger.info("controllo esistenza password");
-//        webTool.waitTime(5);
-//        WebElement messaggio = driver.findElement(By.xpath("//p[contains(text(),'Password:')]"));
-//        getWebDriverWait(10).withMessage("Password non trovato").until(ExpectedConditions.visibilityOf(messaggio));
-//        String password = messaggio.getText().split(": ")[1];
-//        setPassword(password);
         WebElement messaggio = getWebDriverWait(10)
                 .withMessage("Impossibile trovare la Passwrd ")
                 .until(ExpectedConditions.presenceOfElementLocated(By.xpath("//p[contains(text(),'Password:')]")));
@@ -795,12 +707,11 @@ public class HelpdeskPage extends BasePage {
         Path dir = Paths.get(directoryPath);
         try (DirectoryStream<Path> stream = Files.newDirectoryStream(dir)) {
             for (Path entry : stream) {
-                if (Files.isRegularFile(entry)) {
-                    if (extension == null || entry.toString().endsWith(extension)) {
+                if (Files.isRegularFile(entry) && (extension == null || entry.toString().endsWith(extension))) {
                         Files.delete(entry);
                         System.out.println("Deleted file: " + entry.toString());
                     }
-                }
+
             }
         } catch (IOException e) {
             throw new IOException("Failed to delete files in directory: " + directoryPath, e);
